Time: 20 ms (49.07%), Space: 5.9 MB (75.84%) - LeetHub

+++ 9-palindrome-number/9-palindrome-number.cpp
@@ -3,7 +3,7 @@ public:
     bool isPalindrome(int x) {
       if (x < 0) return false;
       
-      long long int reversed = 0;
+      unsigned long int reversed = 0;
       int t = x;
       while (t > 0) {
         reversed *= 10;
